fix hologram issues, cleanup
