+++ Content.Server/_NF/Roles/Systems/InterviewHologramSystem.cs
@@ -8,7 +8,9 @@ using Content.Shared.Humanoid;
 using Content.Shared.Mind;
 using Content.Shared.Mind.Components;
 using Content.Shared.Preferences;
+using Content.Shared.Verbs;
 using Robust.Shared.Player;
+using Robust.Shared.Utility;
 
 namespace Content.Server._NF.Roles.Systems;
 
@@ -21,7 +23,7 @@ public sealed class InterviewHologramSystem : SharedInterviewHologramSystem
     [Dependency] private StationSpawningSystem _stationSpawning = default!;
     [Dependency] private IServerPreferencesManager _prefs = default!;
     [Dependency] private StationSystem _station = default!;
-    [Dependency] private JobTrackingSystem _jobTracking = default!;
+    [Dependency] private StationJobsSystem _stationJobs = default!;
     [Dependency] private ActionsSystem _actions = default!;
 
     public override void Initialize()
@@ -29,6 +31,7 @@ public sealed class InterviewHologramSystem : SharedInterviewHologramSystem
         base.Initialize();
 
         SubscribeLocalEvent<InterviewHologramComponent, MapInitEvent>(OnHologramMapInit);
+        SubscribeLocalEvent<InterviewHologramComponent, GetVerbsEvent<AlternativeVerb>>(OnAlternativeVerb);
         SubscribeLocalEvent<InterviewHologramComponent, MindRemovedMessage>(OnHologramMindRemoved);
         SubscribeLocalEvent<InterviewHologramComponent, MindAddedMessage>(OnHologramMindAdded);
         SubscribeLocalEvent<InterviewHologramComponent, CancelInterviewEvent>(OnHologramCancelInterview);
@@ -41,7 +44,7 @@ public sealed class InterviewHologramSystem : SharedInterviewHologramSystem
         _actions.AddAction(ent, ref ent.Comp.ToggleApprovalActionEntity, ent.Comp.ToggleApprovalAction);
         _actions.SetToggled(ent.Comp.ToggleApprovalActionEntity, ent.Comp.ApplicantApproved);
 
-        // Apply the current character's appearance from their profile, if it exists.
+        // Apply the current character's appearance from their profile if it exists.
         if (!TryComp(ent, out MindContainerComponent? mindContainer)
             || !_mind.TryGetSession(mindContainer.Mind, out var session))
             return;
@@ -49,15 +52,52 @@ public sealed class InterviewHologramSystem : SharedInterviewHologramSystem
         ApplyAppearanceForSession(ent, session);
     }
 
+    // FIXME: This is currently on the server because ShuttleDeed isn't currently properly networked to the client.
+    private void OnAlternativeVerb(Entity<InterviewHologramComponent> ent, ref GetVerbsEvent<AlternativeVerb> ev)
+    {
+        // No access/interact check, should be possible with sight alone
+        if (ev.Hands == null || ev.User == ev.Target)
+            return;
+
+        bool accepted = ent.Comp.CaptainApproved;
+        EntityUid captain = ev.User;
+        bool isCaptain = IsCaptain(ev.User, ent);
+        ev.Verbs.Add(new AlternativeVerb()
+        {
+            Act = () => RaiseLocalEvent(ent, new SetCaptainApprovedEvent(captain, !accepted)),
+            Text = Loc.GetString(accepted ? "interview-hologram-rescind" : "interview-hologram-approve"),
+            Icon = new SpriteSpecifier.Texture(new(accepted ? "/Textures/_NF/Interface/VerbIcons/cross.png" : "/Textures/_NF/Interface/VerbIcons/check.png")),
+            Disabled = !isCaptain,
+            Message = isCaptain ? null : Loc.GetString("interview-hologram-verb-message-need-deed")
+        });
+        ev.Verbs.Add(new AlternativeVerb()
+        {
+            Act = () => RaiseLocalEvent(ent, new DismissInterviewEvent(captain)),
+            Text = Loc.GetString("interview-hologram-dismiss"),
+            Icon = new SpriteSpecifier.Texture(new("/Textures/Interface/VerbIcons/delete.svg.192dpi.png")),
+            Disabled = !isCaptain,
+            Message = isCaptain ? null : Loc.GetString("interview-hologram-verb-message-need-deed"),
+            Priority = -1
+        });
+    }
+
     private void OnHologramMindRemoved(Entity<InterviewHologramComponent> ent, ref MindRemovedMessage ev)
     {
+        // Override job tracking - explicitly reopen the job slot, whatever it was.
+        if (TryComp<JobTrackingComponent>(ent, out var jobTracking))
+        {
+            if (jobTracking.Job != null)
+                _stationJobs.TryAdjustJobSlot(jobTracking.SpawnStation, jobTracking.Job, 1);
+            RemComp<JobTrackingComponent>(ent);
+        }
+
         // Don't let holograms linger.
         QueueDel(ent);
     }
 
     private void OnHologramMindAdded(Entity<InterviewHologramComponent> ent, ref MindAddedMessage ev)
     {
-        // Apply the current character's appearance from their profile, if it exists and hasn't already been applied.
+        // Apply the current character's appearance from their profile if it exists and hasn't already been applied
         if (ent.Comp.AppearanceApplied
             || !TryComp(ent, out MindContainerComponent? mindContainer)
             || !_mind.TryGetSession(mindContainer.Mind, out var session))
@@ -80,6 +120,7 @@ public sealed class InterviewHologramSystem : SharedInterviewHologramSystem
         if (!ent.Comp.ApplicantApproved || !ent.Comp.CaptainApproved)
             return;
 
+        // Entity must have a valid set of coordinates.
         if (!TryComp(ent, out TransformComponent? xform))
             return;
 
@@ -111,9 +152,6 @@ public sealed class InterviewHologramSystem : SharedInterviewHologramSystem
 
     private void OnHologramCancelInterview(Entity<InterviewHologramComponent> ent, ref CancelInterviewEvent ev)
     {
-        if (_mind.TryGetSession(_mind.GetMind(ent), out var session))
-            _gameTicker.Respawn(session);
-
         DismissHologram(ent);
     }
 
@@ -127,12 +165,16 @@ public sealed class InterviewHologramSystem : SharedInterviewHologramSystem
 
     private void DismissHologram(Entity<InterviewHologramComponent> ent)
     {
+        // Override job tracking - explicitly reopen the job slot, whatever it was.
+        if (TryComp<JobTrackingComponent>(ent, out var jobTracking))
+        {
+            if (jobTracking.Job != null)
+                _stationJobs.TryAdjustJobSlot(jobTracking.SpawnStation, jobTracking.Job, 1);
+            RemComp<JobTrackingComponent>(ent);
+        }
+
         if (_mind.TryGetSession(_mind.GetMind(ent), out var session))
             _gameTicker.Respawn(session);
-
-        // Ensure our job slot reopens
-        if (TryComp<JobTrackingComponent>(ent, out var jobTracking))
-            _jobTracking.OpenJob((ent.Owner, jobTracking));
 
         QueueDel(ent);
     }
