+++ Resources/Locale/en-US/_NF/job/hologram.ftl
@@ -4,3 +4,5 @@ interview-hologram-approve = Approve
 interview-hologram-rescind = Rescind Approval
 interview-hologram-dismiss = Dismiss
 interview-hologram-dismiss-self = Dismiss Yourself
+
+interview-hologram-verb-message-need-deed = You need to have an ID with a deed to this ship to do that.
