+++ Resources/Prototypes/_NF/Entities/Mobs/Player/interview_hologram.yml
@@ -1,6 +1,6 @@
 # region Entities
 - type: entity
-  parent: [NFMobBossRestrictions, BaseMob]
+  parent: BaseMob
   id: InterviewHologram
   abstract: true
   components:
@@ -20,9 +20,9 @@
         density: 185
         restitution: 0.0
         mask:
-        - MobMask
+        - SmallMobMask # Pass "through" doors
         layer:
-        - MobLayer
+        - SmallMobLayer # Pass "through" doors
   - type: FloorOcclusion
   - type: Identity
   - type: AnimationPlayer
@@ -38,14 +38,20 @@
   - type: Grammar
     attributes:
       proper: true
-  - type: Tag
-    tags:
-    - DoorBumpOpener
   - type: Inventory
     templateId: hostilehumanoid # Supports loadouts
   - type: InventorySlots
   - type: InteractionOutline
   - type: Clickable
+  - type: MobState # if this is missing, it'll be dead, and ghosting keeps the mind around, which is bad.
+  - type: MobThresholds
+    thresholds:
+      0: Alive
+      100: Critical
+      200: Dead
+  - type: MovementSpeedModifier
+    baseWalkSpeed: 2
+    baseSprintSpeed: 3.5
   # Interview-specific things
   - type: InterviewHologram
     shaderName: Hologram
@@ -56,6 +62,11 @@
     scrollRate: 0.125
   - type: NFSalvageMobRestrictions
     leaveGridPopup: hologram-return-to-shuttle
+    addComponentsLeaveGrid:
+    - type: TimedDespawn
+      lifetime: 10
+    removeComponentsReturnGrid:
+    - type: TimedDespawn
   - type: ActionsContainer
 
 - type: entity
@@ -87,7 +98,7 @@
 - type: entity
   id: ActionInterviewToggleApproval
   name: Toggle Approval
-  description: 
+  description: Toggles whether or not you approve to work with this crew.
   components:
   - type: InstantAction
     icon: _NF/Interface/Actions/acceptOff.png
