+++ Resources/Prototypes/_NF/Maps/debug.yml
@@ -12,11 +12,8 @@
         - type: StationJobs
           availableJobs:
             Contractor: [ -1, -1 ]
-            ContractorInterview: [ -1, -1 ]
             Pilot: [ -1, -1 ]
-            PilotInterview: [ -1, -1 ]
             Mercenary: [ -1, -1 ]
-            MercenaryInterview: [ -1, -1 ]
             Borg: [ -1, -1]
             StationRepresentative: [ -1, -1 ]
             StationTrafficController: [ -1, -1 ]
