+++ Resources/Textures/_NF/Interface/VerbIcons/ATTRIBUTION.txt
@@ -1,0 +1,3 @@
+All icons are public domain.
+
+check and cross pulled from https://github.com/apancik/public-domain-icons.
